% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finfo.R
\name{Dcrit}
\alias{Dcrit}
\title{Determinant of Fisher Information Matrix}
\usage{
Dcrit(w, M, gradient, theta)
}
\arguments{
\item{w}{(\code{numeric}) \cr
Weight vector with length same as rows of \code{M}. Weights must be within 0 and 1
and must sum up to 1.}

\item{M}{(\code{matrix()}) \cr
Design with support points as rows.}

\item{gradient}{(\verb{function(x, theta)})
Gradient function of the model with input parameters \code{x} and \code{theta}.
\code{x} is a \code{numeric()} of length same as number of columns of \code{M}
representing one support point and \cr
\code{theta} is a \code{numeric()} containing the parameters of the model.}

\item{theta}{\code{numeric()} \cr
Parameter vector of the model.}
}
\value{
The determinant of the fisher information matrix.
}
\description{
Can be used as \code{crit} in \link{psoOptDesign}.
}
\examples{
Dcrit(w = c(1/4, 1/4, 1/4, 1/4),
M = rbind(c(1.5, 0.4),
         c(2,   0.3),
         c(5, 0.9),
         c(7, 1)),
gradient = grad_td2pLL_pso,
theta = c(h=2, delta=0.2, gamma=1.3, c0=0.2))
}
