% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcrit_equ_plot_td2pLL.R
\name{dcrit_equ_plot_td2pLL}
\alias{dcrit_equ_plot_td2pLL}
\title{D-optimality equvalence theorem plot for td2pLL model}
\usage{
dcrit_equ_plot_td2pLL(
  des,
  theta,
  time_lim = c(1, 10),
  dose_lim = c(0, 1),
  n_grid = 101,
  return_values = T,
  plot_theta = 45,
  plot_phi = 15,
  plot_r = sqrt(3)
)
}
\arguments{
\item{des}{(\code{list()}) \cr
A calculated design structured as the \link{opt_des_td2pLL} output}

\item{theta}{(\code{numeric(4)}) \cr
Parameter vector of the td2pLL model in the order h, delta, gamma and c0.}

\item{time_lim}{(\code{numeric(2)}) \cr
Lower and upper bound for time}

\item{dose_lim}{(\code{numeric(2)}) \cr
Lower and upper bound of dose}

\item{n_grid}{(\code{integer(1)}) \cr
\code{n_grid} times \code{n_grid} spans the grid for evaluating the equivalence theorem
within the time and dose boundaries}

\item{return_values}{(\code{logical(1)}) \cr
If the equivalence theorem values at each grid point shall be returned,
default is \code{TRUE}.}

\item{plot_theta}{(\code{numeric(1)}) \cr
Theta angle passed to \code{persp}. Default is 45.}

\item{plot_phi}{(\code{numeric(1)}) \cr
Phi angle passed to \code{persp}. Default is 15.}

\item{r_theta}{(\code{numeric(1)}) \cr
Rotation passed to persp`. Default is sqrt(3).}
}
\description{
Using the \link{dcrit_equ_td2pLL} function to calculate the
equivalence theorem inequality at one design points, the correspoinding
equivalence theorem plot can be visualized with this function.
Note that, if the entire surface is equal to or below zero, the design
is D-optimal.
}
\note{
The plot object itself cannot be saved as it is generated with \link{persp}.
}
\examples{
 theta_0 = c(h=2, delta=0.2, gamma=1.3, c0=0.2)
 setwd(1905)
 des <- opt_des_td2pLL(theta =  thtea_0)
 dcrit_equ_plot_td2pLL(des = des, theta = theta_0)
}
