% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcrit_equ_td2pLL.R
\name{dcrit_equ_td2pLL}
\alias{dcrit_equ_td2pLL}
\title{D-optimality equivalence theorem for multidimensional design}
\usage{
dcrit_equ_td2pLL(M_x, des, theta, finfo_inv = NULL)
}
\arguments{
\item{M_x}{(\code{numeric()}) \cr
A single (support) point of the design space}

\item{des}{(\code{list()}) \cr
A calculated design structured as the \link{opt_des_td2pLL} output}

\item{theta}{(\code{numeric(4)}) \cr
Parameter vector of the (non-linear) td2pLL model parameters in the order
h, delta, gamma and c0.}

\item{finfo_inv}{(optimal \code{matrix()})
Inverse of Fisher Information matrix.}
}
\description{
To check if a design is D-optimal, an equivalence theorem
can be used. Precisely, a design \eqn{\xi} is optimal if for all x of the design
space the following holds:
\deqn{\partial \eta(x, \theta)^{\top} M^{-1}(\xi, \theta) \partial \eta(x, \theta) - |\theta| \le 0}
Where \eqn{M} is the Fisher Information matrix of the design assuming
that the td2pLL model with corresponding parameters in \code{theta} is true.
}
