% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finfo.R
\name{finfo}
\alias{finfo}
\title{Fisher Information Matrix for multidimensional designs}
\usage{
finfo(w, M, gradient, theta)
}
\arguments{
\item{w}{(\code{numeric}) \cr
Weight vector with length same as rows of \code{M}. Weights must be within 0 and 1
and must sum up to 1.}

\item{M}{(\code{matrix()}) \cr
Design with support points as rows.}

\item{gradient}{(\verb{function(x, theta)})
Gradient function of the model with input parameters \code{x} and \code{theta}.
\code{x} is a \code{numeric()} of length same as number of columns of \code{M}
representing one support point and \cr
\code{theta} is a \code{numeric()} containing the parameters of the model.}

\item{theta}{\code{numeric()} \cr
Parameter vector of the model.}
}
\value{
Fisher information matrix as a symmetric matrix with \code{length(theta)}
rows.
}
\description{
Fisher Information Matrix for multidimensional designs
}
\details{
Calculates the fisher information matrix for a multidimensional
optimal design in dependency on weights, the design, the gradient and
theta (as we assume a non-linear model).
}
\examples{
finfo(w = c(1/3, 1/3, 1/3),
     M = rbind(c(1.5, 0.4,
               c(2, 0.3),
               c(5, 0.9))),
     gradient = grad_td2pLL_pso,
     theta = c(h=2, delta=0.2, gamma=1.3, c0=0.2))
}
