% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_functions.R
\name{get_EC50s}
\alias{get_EC50s}
\title{Accessing EC50 values of a td2pLL model}
\usage{
get_EC50s(coefs, times)
}
\arguments{
\item{coefs}{Named numeric vector containing the parameters of a td2pLL
model containing the coefficients h, delta, gamma and c0.
For an object of class td2pLL, these can be accessed via \code{coef()}.}

\item{times}{Numeric vector of the times for which the EC50 shall be retrieved.
Entries must be greater than 0.}
}
\value{
A data.frame with columns time and EC50 containing the EC50s at
 \code{times}.
}
\description{
For parameters of a td2pLL model
and given \code{time} values, the EC50 value is retrieved.
}
