% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ED50s.R
\name{get_ED50s}
\alias{get_ED50s}
\title{Accessing ED50 values of a td2pLL model}
\usage{
get_ED50s(coefs, times)
}
\arguments{
\item{coefs}{Named numeric vector containing the parameters of a td2pLL
model containing the coefficients h, delta, gamma and c0.
For an object of class td2pLL, these can be accessed via \code{coef()}.}

\item{times}{Numeric vector of the times for which the ED50 shall be retrieved.
Entries must be greater than 0.}
}
\value{
A data.frame with columns time and ED50 containing the ED50s at
\code{times}.
}
\description{
For parameters of a td2pLL model
and given \code{time} values, the ED50 value is retrieved, i.e.
the dose at which 50 percent of the maximal effect is reached.
}
