% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_starting_values.R
\name{get_starting_values}
\alias{get_starting_values}
\title{Starting values for td2pLL model}
\usage{
get_starting_values(data, h_start = 2, c0_start = 0)
}
\arguments{
\item{data}{data frame with numeric columns named time, dose and resp}

\item{h_start}{optional starting value for the \code{h} parameter.
Default is 2.}

\item{c0_start}{optional starting value for the theshold parameter \code{c0}.
Default is 0.}
}
\value{
List with starting values for h, delta, gamma and c0.
}
\description{
Calculates (cheap) default starting values for fitting a
td2pLL model based on dose time response data.
}
\details{
As starting value for \code{delta} and \code{gamma}, a pair of
cheap starting values of (dose, ED50) at the lowest and the highest
(exposure) time are calculated via the \code{\link[=interp_ED50]{interp_ED50()}} function.
With these two pairs (dose_1, ED50_1) and (dose_2, ED50_2), as
well as the set starting values of \code{h} and \code{c0},
the model equation of the td2pLL model
\deqn{f(d,t)=100-100\frac{d^h}{EC_{50}(t)^h + d^h},}
\deqn{EC_{50}(t) = \Delta \cdot t^{-\gamma} + C_0}
is solved to get the starting values \code{gamma_start} and
\code{delta_start} for \code{gamma} and  \code{delta}.
}
