% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_td2pLL.R
\name{grad_td2pLL}
\alias{grad_td2pLL}
\title{Gradient function of the td2pLL model}
\usage{
grad_td2pLL(time, dose, h, delta, gamma, c0)
}
\arguments{
\item{time}{(\code{numeric(1)})}

\item{dose}{(\code{numeric(1)})}

\item{h}{(\code{numeric(1)})}

\item{delta}{(\code{numeric(1)})}

\item{gamma}{(\code{numeric(1)})}

\item{c0}{(\code{numeric(1)})}
}
\value{
\code{numeric(4)} Gradient vector of the td2pLL model at given
parameters.
}
\description{
For details on the td2pLL model, see \link{fit_td2pLL}
}
