% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_td2pLL.R
\name{grad_td2pLL_pso}
\alias{grad_td2pLL_pso}
\title{Gradient of td2pLL for PSO}
\usage{
grad_td2pLL_pso(x, theta)
}
\description{
Wrapper function of \link{grad_td2pLL} to pass
it to the \link{psoOptDes} function as argument gradient.
}
