% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridOptDesign.R
\name{grid_opt}
\alias{grid_opt}
\title{Optimized weights of td2pLL model (E0 unknown) using grid-search.}
\usage{
grid_opt(
  u_dose,
  u_time,
  iter = 10,
  h,
  delta,
  gamma,
  c0,
  beta_r = 0,
  gif_path = NULL,
  gif_name = NULL
)
}
\arguments{
\item{u_dose}{(\code{numeric()}) \cr
Contains all potential dose points. Consider logarithmic scaling.}

\item{u_time}{(\code{numeric()}) \cr
Contains all potential time points.}

\item{iter}{(\code{integer(1)}) \cr
Number of iterations.}

\item{h}{(\code{numeric(1)})}

\item{delta}{(\code{numeric(1)})}

\item{gamma}{(\code{numeric(1)})}

\item{c0}{(\code{numeric(1)})}

\item{beta_r}{(\code{numeirc(1)}) \cr
Tuning parameter in (0,1). Use 0 (default) for original Titterington.
Some (Dette at al. (2008)) state that \code{beta_r=1} increases convergence speed.}

\item{gif_path}{(\code{character(1)}) \cr
Path were a folder \code{plots} shall be generated and where all plots of
the weights of each iterations shall be saved. No plots generated if NULL.}

\item{gif_name}{(\code{character(1)}) \cr
Directory within \code{gif_path} where a gif from all the plots of the weights per
iteration is saved. Is not generated if \code{iter}>1000 as this likely crashes the
session. No plots or GIF generated if NULL.}
}
\description{
Optimized weights of a td2pLL model with assumed parameters
\code{h}, \code{delta}, \code{gamma} and \code{c0}. Uses multiplicative optimization as
originally described by Titterington (1976).\cr
A less mathematical description is available in Holland-Letz &  Kopp-Schneider (2015).
For the tuning parameter \code{beta_r}, we refer to Dette et al. (2008).
}
\references{
\itemize{
\item Titterington, D. M. (1976). Algorithms for computing D-optimal design
on finite design spaces. In Proc. of the 1976 Conf. on Information Science
and Systems, John Hopkins University (Vol. 3, pp. 213-216).
\item Holland-Letz, T., & Kopp-Schneider, A. (2015). Optimal experimental
designs for dose-response studies with continuous endpoints.
Archives of toxicology, 89(11), 2059-2068. \doi{10.1007/s00204-014-1335-2}
\item Dette, H., Pepelyshev, A., & Zhigljavsky, A. (2008). Improving updating
rules in multiplicative algorithms for computing D-optimal designs.
Computational Statistics & Data Analysis, 53(2), 312-320. \doi{10.1016/j.csda.2008.10.002}
}
}
