% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icaod_info.R
\name{icaod_info}
\alias{icaod_info}
\title{Information matrix for \code{\link[=locally]{locally()}} assuming unknown e0 and known eMax}
\usage{
icaod_info(x, w, param)
}
\arguments{
\item{x}{(\code{numeric(2*7)}) \cr
Contains the current support points under consideration where first all
time coordinates are stored in x and then all dose coordinates.}

\item{w}{(\code{numeric(7)}) \cr
Weights of the support points. \code{sum(w)} must be 1.}

\item{param}{(\code{numeric(5)}) \cr
Contains assumed true parameters for \code{e0}, \code{h}, \code{delta}, \code{gamma}, \code{c0} where
the value of \code{e0} has no effect on the optimal design calculation as it
is a linear parameter.}
}
\value{
Information matrix of dimension 5 times 5.
}
\description{
If e0 is assumed unknown and eMax is assumed known in the td2pLL model,
the optimal design will propose a support point in dose=0, but not in the maximal dose.
To assume additionally that eMax is unknown, use \link{icaod_info_noEmax}.
}
