% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridOptDesign.R
\name{info}
\alias{info}
\title{Information matrix of a td2pLL model with unknown E0}
\usage{
info(u_time, u_dose, w, h, delta, gamma, c0)
}
\arguments{
\item{u_time}{(\code{numeric()}) \cr
Contains all potential time points.}

\item{u_dose}{(\code{numeric()}) \cr
Contains all potential dose points. Consider logarithmic scaling.}

\item{w}{(\code{array(length(u_time), length(u_dose))}) \cr
Weight array. Each row are the weights of the doses for a given time.
All weights must sum up to 1.}

\item{h}{(\code{numeric(1)})}

\item{delta}{(\code{numeric(1)})}

\item{gamma}{(\code{numeric(1)})}

\item{c0}{(\code{numeric(1)})}
}
\value{
(\code{list(2)}) \cr
First element contains derivatives at all points. \cr
Second element contains information matrix.
}
\description{
Used in \link{grid_opt}. E0, i.e. the effect at dose 0
is assumed to be unknown as otherwise the optimal design will
not propose any measurements at dose 0.
}
