% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_ED50.R
\name{interp_ED50}
\alias{interp_ED50}
\title{Starting value calculation for ED50 parameter}
\usage{
interp_ED50(data = NULL)
}
\arguments{
\item{data}{A data.frame with columns dose (numeric) and mean_resp (numeric)
of mean response data at given dose level.}
}
\value{
Returns the numeric \code{ED50_interp}.
}
\description{
Calculates cheap starting parameter for the ED50 parameter of a
2pLL model by interpolation. This function is used in get_starting_values.
}
\details{
The function assumes that the mean response values lay within 0 and
100, i.e. have percent units.
It also assumes a downward trend of the ED50 with dose.
If the lowest mean_resp value is already above 50, the function
returns the maximal dose.
}
