% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_des_td2pLL.R
\name{opt_des_td2pLL}
\alias{opt_des_td2pLL}
\title{Calculate D-optimal design for td2pLL model}
\usage{
opt_des_td2pLL(
  nPoints = 9,
  theta,
  control = NULL,
  Lb = c(1, 0),
  Ub = c(10, 1),
  wFixed = NULL
)
}
\arguments{
\item{nPoints}{(\code{integer()}) \cr
Number of design support points (number of rows of argument \code{M} for other functions).}

\item{theta}{(\code{numeric(4)}) \cr
Parameter vector of the td2pLL model in the order h, delta, gamma and c0.}

\item{control}{(named \code{list()}) \cr
Named list with further arguments for the PSO algorithm: \cr
\code{numit} (\code{integer(1)}) Number of iterations, default is 100 \cr
\code{numPart} (\code{integer(1)}) Number of particles, default is 500 \cr
\code{beta} (\code{numeric(1)}) ???, default is 0.5 \cr
\code{gamma} (\code{numeric(1)}) ???, default is 0.7 \cr
\code{setRsol} (\code{logical(1)}) ??, default is FALSE \cr
\code{setProgressBar} (\code{logical(1)}) If a progress bar shall be printed,
default is FALSE \cr
\code{OutCritValue} (\code{logical(1)}) If the final value of the criterion shall be
added to the output, default is FALSE. \cr
\code{intmRes} (logical(1)) If intermediate results shall be ???, default is FALSE \cr}

\item{Lb}{(\code{numeric()}) \cr
Lower bounds for the support points (e.g. for time and dose in td2pLL).}

\item{Ub}{(\code{numeric()}) \cr
Upper bound for the support points (used e.g. for time and dose in td2pLL).}

\item{wFixed}{(Optional numeric(\code{nPoints})) \cr
(Fixed) weights. Must have length \code{nPoints} and weights must sum up to 1.
Giving equal weights might improve the performance of the algorithm.}
}
\description{
Calculate a (local) D-optimal design for a td2pLL model using
particle swarm optimization.
}
\examples{
res <- opt_des_td2pLL(theta =  c(h=2, delta=0.2, gamma=1.3, c0=0.2))
res
}
