% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_td2pLL.R
\name{plot.td2pLL}
\alias{plot.td2pLL}
\title{Plot interactive td2pLL models}
\usage{
\method{plot}{td2pLL}(
  td2pLL_model = NULL,
  td2pLL_coefs = NULL,
  dose_lim = c(1e-04, 1),
  time_lim = c(1, 7),
  add_data = NULL,
  n_grid = 100,
  title = NULL,
  xaxis_scale = "log",
  yaxis_scale = "-",
  xaxis_title = "Dose",
  yaxis_title = "Time",
  zaxis_title = "Response",
  add_ED50_line = TRUE,
  ED50_line_col = "red",
  ED50_line_width = 6,
  ...
)
}
\arguments{
\item{td2pLL_model}{A td2pLL object generatet via \code{fit_td2pLL}. If not
provided, alternatively, \code{td2pLL_coefs} can be provided.}

\item{td2pLL_coefs}{Named numeric vector with parameters for h, delta, gamma and c0
of the td2pLL model.}

\item{dose_lim}{Boundaries of the doses (xaxis) to be plotted. Default is
\code{c(1e-4, 1)}.
Note: If \code{xaxis_scale = "log"} (default), then \code{dose_lim} cannot include 0.
If \code{dose_lim} shall include the 0, set \code{xaxis_scale = "linear"}.}

\item{time_lim}{Boundaries for the time (xaxis). Default is \code{c(1, 7)}.}

\item{add_data}{Optional \code{data.frame} to add data points to the
surface plot. Must include columns \code{dose}, \code{time} and \code{resp}.
n_grid Integer. \code{n_grid*n_grid} is the grid for surface evaluations
that will be interpolated. Increase \code{n_grid} for smoother plot.}

\item{title}{Optional plot title.}

\item{xaxis_scale}{Scale of x-axis (Dose-Axis). Character of c("log", "linear", "-").
If "-" is set, then plot_ly tries to guess which scale shall be used.}

\item{yaxis_scale}{Scale of y-axis (Time-Axis). Character of c("log", "linear", "-").
If "-" is set, then plot_ly tries to guess which scale shall be used.}

\item{xaxis_title}{Title for Dose-Axis.}

\item{yaxis_title}{Title for Time-Axis.}

\item{zaxis_title}{Title for Response-Axis.}

\item{add_ED50_line}{Logical. Indicates if the line of EC_50 values shall be
annotated.}

\item{ED50_line_col}{Color for optionally added ED_50 line.}

\item{ED50_line_width}{Width for optionally added ED_50 line.}
}
\description{
\code{plot.td2pLL} is the plot method for the S3 class
\code{td2pLL}.
Model fits generated with the fit_td2pLL functions are of class
\code{c("td2pLL", "nls")} and can therefore be used for this plot method.
If no fitted model but a through parameters pre-specified td2pLL model
shall be plotted, this can be done via the \code{td2pLL_coefs} argument.
For details on the \code{td2pLL} model, see \code{\link{fit_td2pLL}}.
If the \code{TDR} function is used which performs the two-step
modeling pipeline, one can apply plot.td2pLL to the \code{fit} list entry
returned by TDR, if fitting a \code{td2pLL} model was chosen in accordance
to the anova pre-test.
}
\details{
For further details on the td2pLL model, check \code{\link{fit_td2pLL}}.
For details on the ANOVA used, see \code{\link{td2pLL_anova}}. More over,
the entire procedure is explained in duda et al. (2021).
For plotting, the \code{plot_ly} function of package \code{plotly} is used.
}
