% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_td_dcrit_equ.R
\name{plot_td_dcrit_equ}
\alias{plot_td_dcrit_equ}
\title{Plot sensitivity function of calculated optimal design to verify optimality}
\usage{
plot_td_dcrit_equ(
  td_opt_des,
  time_lim = NULL,
  dose_lim = NULL,
  n_grid = 61,
  title = NULL,
  return_values = FALSE,
  plot_theta = 120,
  plot_phi = 30,
  plot_r = sqrt(3)
)
}
\arguments{
\item{td_opt_des}{Optimal design for td2pLL model calculated with \link{td_opt}}

\item{time_lim}{(\code{numeric(2)}) \cr
Lower and upper bound of the time component.}

\item{dose_lim}{(\code{numeric(2)}) \cr
Lower and upper bound of the dose component.}

\item{n_grid}{(\code{numeric(1)}) \cr
The values will be calculates on a \code{n_grid} time \code{n_grid} grid.}

\item{title}{(\code{character(1)}) \cr
Optional plot title. By default, the maximal value and its position are
displayed in the title.}

\item{return_values}{(\code{logical(1)}) \cr
Defines if the values of the points are returned.}

\item{plot_theta, plot_phi, plot_r}{Angles theta, phi and r passed to \link{persp}}
}
\description{
The values at each point (time, dose) is the gradient at the point
times the inverse of the FIM of the calculated optimal design (at the assumes
parameters) times the gradient.
If the resulting surface is below zero, the design is optimal and in the support
points the value is zero. \cr
Due to computational inaccuracies, the maximal value is usually not exactly zero.
}
