% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_td_des.R
\name{plot_td_des}
\alias{plot_td_des}
\title{Visualize design calculated with \link{td_opt}}
\usage{
plot_td_des(td_opt_des, label_weights = T, log10_dose = T)
}
\arguments{
\item{td_opt_des}{Optimal design for td2pLL model calculated with \link{td_opt}}

\item{label_weights}{(\code{logical(1)}) \cr
Add labels with weights to the support points.}

\item{log10_dose}{(\code{logical(1)}) \cr
If \code{TRUE} (default), the dose is displayed on a log10 scale.}
}
\description{
Visualize design calculated with \link{td_opt}
}
