% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psoOptDesign.R
\name{psoOptDesign}
\alias{psoOptDesign}
\title{PSO algorithm for multidimensional optimal designs}
\usage{
psoOptDesign(
  crit,
  control = list(),
  nPoints = 3,
  dimension = 1,
  Lb = 0,
  Ub = 150,
  xFixed = NULL,
  wFixed = NULL,
  xold = NULL,
  nold = rep(0, length(xold)),
  n2 = rep(0, length(old)),
  ...
)
}
\arguments{
\item{crit}{(\code{function})cr
Function that calcultion the optimality criterion of interest.
I.e. for D-optimality the determinant.
The \code{crit} funciton must have parameters:
\code{M} (matrix) with support points as rows,
\code{w} (vector) with weights,
\code{gradient} a function for the model's gradient depending on
\code{x} (numeric(\code{dimension})) which is a single point of the design and
\code{theta} (numeric()) which contains the model parameters at which the
gradient shall be evaluated.}

\item{control}{(named \code{list()}) \cr
Named list with further arguments for the PSO algorithm: \cr
\code{numIt} (\code{integer(1)}) Number of iterations, default is 100 \cr
\code{numPart} (\code{integer(1)}) Number of particles, default is 300 \cr
\code{beta} (\code{numeric(1)}) ???, default is 0.5 \cr
\code{gamma} (\code{numeric(1)}) ???, default is 0.7 \cr
\code{setRsol} (\code{logical(1)}) ??, default is FALSE \cr
\code{setProgressBar} (\code{logical(1)}) If a progress bar shall be printed,
default is FALSE \cr
\code{OutCritValue} (\code{logical(1)}) If the final value of the criterion shall be
added to the output, default is FALSE. \cr
\code{intmRes} (logical(1)) If intermediate results shall be ???, default is FALSE \cr}

\item{nPoints}{(\code{integer()}) \cr
Number of design support points (number of rows of argument \code{M} for other functions).}

\item{dimension}{(\code{integer()}) \cr
Number of dimensions for one support point (number of columns of argument \code{M}
for other functions). Default is 1.}

\item{Lb}{(\code{numeric()}) \cr
Lower bounds for the support points (e.g. for time and dose in td2pLL).}

\item{Ub}{(\code{numeric()}) \cr
Upper bound for the support points (used e.g. for time and dose in td2pLL).}

\item{xFixed}{(Optional numeric \code{matrix}) \cr
(Fixed) support Points, default is NULL. One point is one row. \cr}

\item{wFixed}{(Optional numeric(\code{nPoints})) \cr
(Fixed) weights. Must have length \code{nPoints} and weights must sum up to 1.
Giving equal weights might improve the performance of the algorithm.}

\item{xold}{???}

\item{nold}{???}

\item{n2}{???}

\item{...}{???}
}
\description{
Calculate an optimal design for a multi-dimensional
model, e.g. the td2pLL \link{fit_td2pLL} model, using particle swarm optimization
(PSO) algorithm.
Code kindly provided by Kirsten Schorning.
}
