% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td2pLL_anova.R
\name{td2pLL_anova}
\alias{td2pLL_anova}
\title{ANOVA pre-test in the TDR analysis pipeline}
\usage{
td2pLL_anova(data, alpha = 0.05)
}
\arguments{
\item{data}{(\code{data.frame()})\cr
Numeric data frame containing the response \code{resp}, \code{time} and \code{dose}.}

\item{alpha}{(\code{numeric(1)})\cr
The significance level.}
}
\description{
In the time-dose-response (TDR) analysis pipeline,
data where in addition to dose (or concentration) and response, also
time (e.g. exposure times) are measured, it is checked in a (nested) anova
based pre-test if the td2pLL model is appropriate.
That means, it is checked if the exposure time has an effect on the
dose-response relationship or not.
}
\details{
A nested anova is performed where a null model is tested against
a full model. The null model is a regular 2pLL model with upper and
lower limit 100 and 0, respectively.
The (exposure) time variable is ignored and a single dose-response
curve is fitted for the null model.
In the full model, an individual ED50 value is calculated for each (exposure)
time level. However, only one common \code{h} parameter is included in
the full model.
Hence, if the test rejects the null model, there seems to be significant
evidence that (exposure) time has an effect on the dose-response
relationship. Thus, the td2pLL model will be used during the fitting step
of the TDR analysis pipeline. Otherwise, if there is no significant evidence
of the influence of (exposure) time on the dose-response relationship,
a simple 2pLL model is used in the fitting step that ignores (exposure)
time.
For more details see Duda et al. (2021).
}
\examples{
data(cytotox)
data_subset <- cytotox[cytotox$compound == "ASP", c("expo", "dose", "resp")]
colnames(data_subset)[1] <- "time"
td2pLL_anova(data = data_subset)
}
