% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td2pLL_plot_helpers.R
\name{td2pLL_plot_add_data}
\alias{td2pLL_plot_add_data}
\title{helper function for plotting td2pLL model}
\usage{
td2pLL_plot_add_data(plotly_plot, data_to_add)
}
\arguments{
\item{plotly_plot}{\cr
A plotly plot generated with \link{td2pLL_plot_basis}}

\item{data_to_add}{(\code{data.frame})\cr
Numeric data frame with columns \code{dose}, \code{time} and \code{resp}.}
}
\description{
Adds data points to plot
}
