% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td2pLL_plot_helpers.R
\name{td2pLL_plot_add_ED50}
\alias{td2pLL_plot_add_ED50}
\title{helper function for plotting td2pLL models}
\usage{
td2pLL_plot_add_ED50(
  plotly_plot,
  td2pLL_coefs,
  time_seq,
  dose_lim,
  ED50_line_col,
  ED50_line_width
)
}
\arguments{
\item{plotly_plot}{\cr
A plotly plot generated with \link{td2pLL_plot_basis}}

\item{td2pLL_coefs}{(named \code{numeric(4)})\cr
Contains parameters for \code{h}, \code{delta}, \code{gamma} and \code{c0}
of the \link{td2pLL} model.}

\item{time_seq}{(\code{numeric()})}

\item{dose_lim}{(\code{numeric(2)})\cr
Boundaries of the doses (xaxis) for plotting.
Note: If \code{xaxis_scale = "log"} (default), then \code{dose_lim} cannot include 0.
If \code{dose_lim} shall include the 0, set \code{xaxis_scale = "linear"}.}

\item{ED50_line_col}{(\code{character(1)})\cr
Color for optionally added ED50 line.}

\item{ED50_line_width}{(\code{numeric(1)})\cr
Width for optionally added ED50 line.}
}
\description{
Adds ED50 line to td2pLL model plot
}
