% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td2pLL_plot_helpers.R
\name{td2pLL_plot_basis}
\alias{td2pLL_plot_basis}
\title{helper function for plotting td2pLL model}
\usage{
td2pLL_plot_basis(
  dose_seq,
  time_seq,
  input_grid,
  xaxis_title,
  xaxis_scale,
  yaxis_title,
  yaxis_scale,
  zaxis_title,
  title
)
}
\arguments{
\item{dose_seq}{(\code{numeric()})}

\item{time_seq}{(\code{numeric()})}

\item{input_grid}{(\code{data.frame()})\cr
Contains the function values. Number of rows is length of \code{dose_seq},
number of cols is length of \code{time_seq}. The (i,j)th element is the
response value at dose \code{dose_seq}(i) and \code{time_seq}(j).}

\item{xaxis_title}{(\code{character(1)})\cr
Title for dose-axis, time-axis and response-axis.}

\item{xaxis_scale}{(\code{character(1)} in \code{c("log", "linear", "-")}\cr
Scale of x-axis (dose-axis).\cr
If \code{"-"} is set, then \code{\link[=plot_ly]{plot_ly()}} tries to guess which scale to use.}

\item{yaxis_title}{(\code{character(1)})\cr
Title for dose-axis, time-axis and response-axis.}

\item{yaxis_scale}{(\code{character(1)} in \code{c("log", "linear", "-")}\cr
Scale of y-axis (time-axis).\cr
If \code{"-"} is set, then \code{\link[=plot_ly]{plot_ly()}} tries to guess which scale to use.}

\item{zaxis_title}{(\code{character(1)})\cr
Title for dose-axis, time-axis and response-axis.}

\item{title}{(\code{character(1)})\cr
Optional plot title.}
}
\description{
Generates basis of td2pLL plot in plotly
}
