% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td2pLL_plot_helpers.R
\name{td2pLL_plot_title}
\alias{td2pLL_plot_title}
\title{helper function for plotting td2pLL model}
\usage{
td2pLL_plot_title(td2pLL_coefs)
}
\arguments{
\item{td2pLL_coefs}{(named \code{numeric(4)})\cr
Contains parameters for \code{h}, \code{delta}, \code{gamma} and \code{c0}
of the \link{td2pLL} model.}
}
\value{
\code{character(1)}
}
\description{
Generates title of plot based on coefficients
}
