% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icaodOptDesign.R
\name{td_opt}
\alias{td_opt}
\title{Calculate optimal locally D-optimal design for td2pLL model}
\usage{
td_opt(
  param,
  Emax_known = F,
  lx = c(1, exp(-8)),
  ux = c(10, 1),
  iter = 1000,
  ICA.control = NULL
)
}
\arguments{
\item{param}{(\code{numeric(4)}) \cr
Assumed parameters for h, delta, gamma and c0 in this order.}

\item{Emax_known}{(\code{logical(1)}) \cr
Whether or not the maximal effect is known and hence, if it is required
to test for larger doses.
If \code{FALSE} (default) then the maximal effect (which is -100 in a td2pLL model
for typical cytotoxicity data) is assumed unknown and hence the optimal design
will propose to make measurement at the largest possible dose in order to
gather information on the unknown, maximal effect \code{Emax}. \cr
Note that the mean control level \code{e0} which is assumed \code{e0=100} in the
typical td2pLL model is, for optima design planning, always assumed to be
unknown as otherwise no control measurements in dose=0 would be porposed
by the optimal design.
@param lx (\code{numeric(2)}) \cr
Minimal value for time and dose. Note that for dose, you cannot give
a lower bound of zero due to technical reasons.
Instead, provide a very small lower bound (default is exp(-8) = 0.00034)
that represents zero.
@param ux (\code{numeric(2)}) \cr
Maximal value for time and dose.
@param iter (\code{integer(1)}) \cr
Number of iterations in the Imperialist Competitive Algorithm. Default is 1000.
@param ICA.control(\code{list()}) \cr
Additional control parameters passed to the \link[ICAOD:locally]{ICAOD::locally} function.
See \code{\link[ICAOD:ICA.control]{ICAOD::ICA.control()}} for details.
@return An object of class \code{minimax} generated with \code{\link[ICAOD:locally]{ICAOD::locally()}}.
For \code{Emax_known} = \code{TRUE}, the design has 7 support points. Otherwise,
it has 8.
@examples
td_opt_1 <- td_opt(param = c(h = 2, delta = 0.2, gamma = 1.3, c0 = 0.2),
Emax_known = T, ICA.control = c(ncount = 300, rseed = 1905))
plot_td_des(td_opt_1)
plot_td_dcrit_equ(td_opt_2)
td_opt_2 <- td_opt(param = c(h = 2, delta = 0.2, gamma = 1.3, c0 = 0.2),
Emax_known = F, ICA.control = c(ncount = 300, rseed = 1905))}
}
\description{
If prior knowledge on the (rough) model shape is available,
an optimal experimental design for a td2pLL model can be calculated.
The numerical back bone is the Imperialist Competitive Algorithm
implemented by Masoudi et al (2017).
}
