% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR.R
\name{TDR}
\alias{TDR}
\title{Time-Dose-Response analysis pipeline}
\usage{
TDR(data, alpha = 0.05, strict_stop = FALSE, ...)
}
\arguments{
\item{data}{(Numeric \code{data.frame()})\cr
Data frame with columns named \code{time}, \code{dose} and \code{resp}.\cr
Note that the data is expected to be on the percent scale with values
(roughly) within 0 and 100.}

\item{alpha}{(\code{numeric(1)}in (0,1))\cr
1- alpha is the confidence level for testing in step 1.}

\item{strict_stop}{(\code{logical(1)})\cr
Optional logical. When \code{FALSE}, the default, then
in case of an error due to non-convergence in the pre-test, then in the
second step a simple 2pLL model is fitted as if the pre-test was non-
significant.
If \code{strict_stop} is \code{TRUE} and there is an error due to
non-convergence in the pre-test, the procedure stops and no model is fitted
in step 2.}

\item{...}{Further arguments that can be passed on to \code{\link[=fit_td2pLL]{fit_td2pLL()}}.}
}
\value{
A list with entries \code{pretest} and \code{fit}.
\code{pretest} is captures the anova based pre-test result as a list with
entires \code{signif} (TRUE/FALSE or NA if no-convergence), \code{alpha},
\code{anova_res} (the anova result from function \cite{anova}) and
\code{conv} (logical: If the pre-test converged).
\code{fit} Is, depending on the pre-test, either an object of class
\code{td2pLL} or a 2pLL fit, i.e. an object of class \code{drc}.
}
\description{
\code{TDR} performs the time-dose-response analysis pipeline as
presented in Duda et al. (2021). That is: For a dose-response or
concentration-response data set where, additionally also the (exposure)
time is varied, this procedure can be applied. The main aim of this
procedure are cytotoxicity data.
The approach is divided into two steps. In step one, it is tested in a
nested ANOVA if the (exposure) time has an influence on the dose-response
relationship. In case of a significant results, the time-dose two-parameter
log-logistic model is fitted to the data:
\deqn{f(d,t)=100-100\frac{d^h}{EC_{50}(t)^h + d^h}}
with
\deqn{EC_{50}(t) = \Delta \cdot t^{-\gamma} + C_0.}
If no significant result is obtained, a dose-response 2pLL curve is fitted,
ignoring the information on (exposure) time.
}
\details{
For further details on the td2pLL model, check \code{\link[=fit_td2pLL]{fit_td2pLL()}}.
For details on the ANOVA used, see \code{\link[=td2pLL_anova]{td2pLL_anova()}}. More over,
the entire procedure is explained in duda et al. (2021).
}
\examples{
data(cytotox)
data_subset <- cytotox[cytotox$compound == "ASP", c("expo", "dose", "resp")]
colnames(data_subset)[1] <- "time"
TDR_res <- TDR(data = data_subset)
# Pre-test rejected time dependency, so a regular 2pLL model is the result
plot(TDR_res$fit)
data_subset <- cytotox[cytotox$compound == "CHL", c("expo", "dose", "resp")]
colnames(data_subset)[1] <- "time"
TDR_res <- TDR(data = data_subset)
# Pre-test did not reject time dependency, so a  td2pLL model is the result
# Note that the interactive Plot is in the Viewer panel, not in the Plots panel
plot(TDR_res$fit)
}
