% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_fitting.R
\name{fit_joint_2pLL}
\alias{fit_joint_2pLL}
\title{Fitting a 2pLL model}
\usage{
fit_joint_2pLL(data)
}
\arguments{
\item{data}{data frame containing the dose and resp variable}
}
\value{
An object of class \code{drc}.
}
\description{
\code{fit_joint_2pLL} fits a 2pLL model. The "joint" is in the
name because, despite being a regular 2pLL model, this function is used
in the anova-td2pLL pipeline \code{\link{TDR}}: If the anova pre-test, that
checks if there is a difference in EC50 parameters between (exposure) times,
is not significant, the exposure times are ignored. In other words,
a single, 'joint', dose-response curve is fitted to the data,
ignoring the information of exposure time.
}
\details{
The function is a wrapper function using the LL2.2() argument
from the \code{\link[drc]{drm}} function with the fixed asymptotes
upper=100 and lower=0.
As a first try, the BFGS method is used as optimization procedure.
If this yields an error, the Nelder-Mead method is used as opitmization
method.
}
\note{
When using the LL2.2() model from the \code{\link[drc]{drm}} function,
the EC50 parameter is parametrized as log(EC50). We do this for improved
stability.
}
