% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ED50s.R
\name{get_ED50s}
\alias{get_ED50s}
\title{Accessing ED50 values of a td2pLL model}
\usage{
get_ED50s(coefs, times)
}
\arguments{
\item{coefs}{(Named \code{numeric(4)})\cr
Contains the parameters of a td2pLL, namely \code{h}, \code{delta}, \code{gamma} and \code{c0}.
For an object of class \code{td2pLL}, these can be accessed via \code{coef()}.}

\item{times}{(\code{numeric()})\cr
Times at which the ED50 shall be retrieved.
Entries must be greater than 0.}
}
\value{
(\code{data.frame()})\cr
A data.frame with columns \code{time} and \code{ED50}.
}
\description{
For parameters of a td2pLL model
and given \code{time} values, the ED50 value is retrieved, i.e.
the dose at which 50 percent of the maximal effect is reached.
}
