% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_td2pLL.R
\name{plot.td2pLL}
\alias{plot.td2pLL}
\title{Plot interactive td2pLL models}
\usage{
\method{plot}{td2pLL}(
  td2pLL_model = NULL,
  td2pLL_coefs = NULL,
  dose_lim = c(1e-04, 1),
  time_lim = c(1, 7),
  add_data = NULL,
  n_grid = 100,
  title = NULL,
  xaxis_scale = "log",
  yaxis_scale = "-",
  xaxis_title = "Dose",
  yaxis_title = "Time",
  zaxis_title = "Response",
  add_ED50_line = T,
  ED50line_col = "red",
  ED50line_width = 6
)
}
\description{
\code{plot.td2pLL} is the plot method for the S3 class
\code{td2pLL}.
 Model fits generated with the fit_td2pLL functions are of class
 \code{c("td2pLL", "nls")} and can therefore be used for this plot method.
 If no fitted model but a through parameters pre-specified td2pLL model
 shall be plotted, this can be done via the \code{td2pLL_coefs} argument.
 For details on the \code{td2pLL} model, see \code{\link{fit_td2pLL}}.
}
