% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_fitting.R
\name{td2pLL}
\alias{td2pLL}
\title{Calculate response value for given td2pLL model}
\usage{
td2pLL(time, dose, h, gamma, c0, delta)
}
\arguments{
\item{time}{numeric for the time value where the response shall be calculated}

\item{dose}{numeric for the dose (or concentration) value where the response
shall be calculated}

\item{h}{The \code{h} parameter of the model}

\item{gamma}{The \code{gamma} parameter of the model}

\item{c0}{The \code{c0} parameter of the model}

\item{delta}{The \code{delta} parameter of the model}
}
\value{
The response value of the model \emph{in percent} at the given \code{time} and
\code{dose} value.
}
\description{
\code{td2pLL} returns the response value of a fully specified
time-dose two-parameter log-logistic model at a certain time and dose
value:
\deqn{f(d,t)=100-100\frac{d^h}{EC_{50}(t)^h + d^h}}
with
\deqn{EC_{50}(t) = \Delta \cdot t^{-\gamma} + C_0}
}
\examples{
td2pLL(time = 4, dose = 0.1, h = 2, gamma = 2.5, c0 = 0.1, delta = 0.3)
}
